Review: Change config structure
