+++ Remotion/ObjectBinding/Web.Development.WebTesting.IntegrationTests/App.config
@@ -27,8 +27,10 @@
       requestErrorDetectionStrategy="AspNet">
     <hosting name="IisExpress" type="IisExpress" port="60402" />
     <testsite path=".\..\..\..\Web.Development.WebTesting.TestSite">
-      <resource path="..\..\Web\Core\res" />
-      <resource path="..\..\ObjectBinding\Web\res" />
+      <resources>
+        <add path="..\..\Web\Core\res" />
+        <add path="..\..\ObjectBinding\Web\res" />
+      </resources>
     </testsite>
     <chrome disableSecurityWarningsBehavior="Ignore" />
     <edge disableSecurityWarningsBehavior="Ignore" />
