+++ Remotion/Web/Development.WebTesting.IntegrationTests/App.config
@@ -30,7 +30,9 @@
       requestErrorDetectionStrategy="Remotion.Web.Development.WebTesting.IntegrationTests.DiagnosticInformationCollectioningRequestErrorDetectionStrategy, Remotion.Web.Development.WebTesting.IntegrationTests">
     <hosting name="IisExpress" type="IisExpress" port="60401" />
     <testsite path=".\..\..\..\Development.WebTesting.TestSite">
-      <resource path="..\..\Web\Core\res" />
+      <resources>
+        <add path="..\..\Web\Core\res" />
+      </resources>
     </testsite>
     <chrome disableSecurityWarningsBehavior="Ignore" />
     <edge disableSecurityWarningsBehavior="Ignore" />
