+++ Remotion/Web/Development.WebTesting/Schemas/WebTestingConfiguration.xsd
@@ -18,13 +18,19 @@
         </xs:element>
         <xs:element name="testsite" minOccurs="0" maxOccurs="1">
           <xs:complexType>
-            <xs:sequence minOccurs="0">
-              <xs:element name="resource">
+            <xs:all>
+              <xs:element name="resources" minOccurs="0" maxOccurs="1">
                 <xs:complexType>
-                  <xs:attribute name="path" type="c:RequiredStringType" use="required" />
+                  <xs:sequence>
+                    <xs:element name="add">
+                      <xs:complexType>
+                        <xs:attribute name="path" type="c:RequiredStringType" use="required" />
+                      </xs:complexType>
+                    </xs:element>
+                  </xs:sequence>
                 </xs:complexType>
               </xs:element>
-            </xs:sequence>
+            </xs:all>
             <xs:attribute name="path" type="c:RequiredStringType" use="required" />
           </xs:complexType>
         </xs:element>
