+++ Remotion/Web/Development.WebTesting/Web.Development.WebTesting.csproj
@@ -102,6 +102,7 @@
     <Compile Include="Configuration\DriverConfiguration.cs" />
     <Compile Include="Configuration\TestSiteConfigurationElement.cs" />
     <Compile Include="Configuration\TestSiteResourceConfigurationElement.cs" />
+    <Compile Include="Configuration\TestSiteResourceConfigurationElementCollection.cs" />
     <Compile Include="DriverConfigurationOverride.cs" />
     <Compile Include="BrowserSession\BrowserLogEntry.cs" />
     <Compile Include="BrowserSession\Edge\EdgeBrowserSession.cs" />
